+++ lib/main.dart
@@ -4,7 +4,7 @@ import 'dart:io';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:rpmlauncher/handler/window_handler.dart';
-import 'package:rpmlauncher/screen/LauncherHome.dart';
+import 'package:rpmlauncher/screen/main_screen.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:system_info/system_info.dart';
 import 'package:rpmlauncher/model/account/Account.dart';
@@ -116,7 +116,7 @@ Future<void> run() async {
       },
     );
 
-    runApp(const LauncherHome());
+    runApp(const MainScreen());
 
     logger.info("OS Version: ${await RPMLauncherPlugin.platformVersion}");
 
@@ -124,7 +124,7 @@ Future<void> run() async {
       await WindowHandler.setFullScreen(true);
     }
 
-    await googleAnalytics.ping();
+    await googleAnalytics?.ping();
 
     logger.info("Start Done");
   }, (exception, stackTrace) async {
