+++ pubspec.yaml
@@ -12,12 +12,14 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  flutter_localizations:
+    sdk: flutter
   http: ^0.13.4
   url_launcher: ^6.0.18
   path_provider: ^2.0.1
-  split_view: ^3.1.0 
+  split_view: ^3.1.0
   path: ^1.8.0
-  archive: ^3.1.8
+  archive: ^3.3.0
   crypto: ^3.0.1
   system_info: ^1.0.1
   oauth2: ^2.0.0
