+++ test/launcher_home_test.dart
@@ -3,20 +3,16 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:rpmlauncher/screen/HomePage.dart';
-import 'package:rpmlauncher/screen/LauncherHome.dart';
+import 'package:rpmlauncher/screen/main_screen.dart';
 
 import 'util/test_util.dart';
 
 void main() {
   setUpAll(() => TestUtil.init());
   testWidgets('Launcher Home', (WidgetTester tester) async {
-    await tester.runAsync(() async {
-      await tester.pumpWidget(const MaterialApp(home: LauncherHome()));
-    });
+    await tester.pumpWidget(const MainScreen());
   });
   testWidgets('Home Page', (WidgetTester tester) async {
-    await tester.runAsync(() async {
-      await tester.pumpWidget(const MaterialApp(home: HomePage()));
-    });
+    await TestUtil.baseTestWidget(tester, const HomePage(), async: true);
   }, skip: Platform.isMacOS);
 }
